Moved NotPromotableError back to IotaAPI
